Create simple memory test in C
This test is useful when making changes to the memory sub-system. It
first tests writing data to RAM and reading it back. Then it writes to a
peripheral register and checks the value that is read back. At the
moment it passes and fails using an infinite while loop so that you can
check the result using a wave output from simulation.