Add support for Sonata PCB
This commit assumes revision 0.2 of the board. It adds a pin file, a top
level module, a clock generator and the appropriate changes to the core
file. It also updates the getting started guide to use the Sonata PCB by
default.

This commit is derived from a commit in the demo system:
https://github.com/lowRISC/ibex-demo-system/commit/d9ec0d001d93585cbecf94a6c43de755f3cbbd34
With the following changes:
- Remove the JTAG pins from the top and the XDC file since Sonata does
not have a debug module yet.
- Rename the pins to match the names in the PCB documentation.