[doc] Update memory map for UART, I2C and SPI
Initially, each instance of these blocks had 0x100 amount of memory
which is enough for 64 32-bit registers. To avoid any problem in the
future I increased the memory by each block to 0x1000. I also increase
the distance between the UART, I2C and SPI regions in the main memory
map so that the maximum number of blocks possible stays the same.