Timer test code
Test for code rv_timer. This is built as C++ code in
the same manner as the existing bare metal C code,
but is intended to migrate to the CHERIoT code base
at some point.
In particular note that constructors and destructors
of global statics are not invoked by this temporary
build setup.