[rtl] More thorough LED clearing on power up
It's been observed that the existing LED clearing on power up was not
sufficient as the LEDs were turning on on FPGA reset.

This adds a time deley before we first clear the LEDs as well as a
repeated clear to ensure it takes effect.

Additionally some comments have been added to the existing RTL.