Reset SPI flash before accessing it
If the system is reset while a SPI access is going on, the SPI device
might not be in the correct state and using it directly will fail,
and we don't show helpful message in this case, just
"Failed ELF Magic Check".

Add a software reset sequence to reset it to a well known state before
accessing.