Add XADC
Add a generated wrapper for the XADC built-in to the FPGA
and an adapter to connect it to the TLUL crossbar
via the XADC's Dynamic Reconfiguration Port (DRP).

The DRP registers are all 16-bits, and have been mapped
at 32-bit intervals in the TLUL address space.
All writes must be 16-bit writes or they will be silently dropped.
Any writes outside of the range of writable registers will be
silently dropped.
Any reads above the range of DRP registers will cause an error.