[doc] Fix getting started link in top-level README
Now the getting started guide lives in a 'README.md' we cannot link the
README.md directly as it will work on github but not in the rendered
mdbook. Instead link to the directory containing the getting started
guide.

On github this still takes you to a page with the rendered
documentationm just with a directory listing on top of it. Crucially it
fixes the rendered mdbook version.