Implement SPI DPI and models of ST7735 LCD and W25Q256 flash
DPI implementation of SPI device. SystemVerilog module is
parameterised for multiple chip selects, devices and out-of-
band signals in each direction.
Sampling occurs on SCK rising edge, launching of data on
SCK falling edge.
Software model of basic ST7735 LCD functionality produces a
PPM file showing the image displayed on the LCD.
Implement a simple model of the flash device that we use on
the Sonata FPGA; sufficient for the `spi_test` code to produce
the same output as on the physical system.