Create basic USB device test based on existing check
This commit incorporates USB device tests into the existing Sonata Cheri
test runner, and for now just adds a simple smoke/configuration test
based on the existing USB device check that simply attempts to connect
and set up a basic configuration.

This also makes a minor change to the hyperram test to avoid name
clashes on defines used to specify the number of iterations of each
test.