Update CHERI SPI flash utils to use 4-byte addresses
Currently most of the SPI flash utils driver is designed around the use
of 3-byte address commands. However, the flash being used actually is
larger in size and supports some small 4-byte addresses.

To allow the writing of pages and the erasing of sectors within the
entirity of flash (for testing or other purposes), this commit changes
the page write and sector erase commands to their 4-address equivalents.