Add CHERI SPI test
This commits adds a SPI test to the existing baremetal CHERI test
runner. For now, this is a simple smoketest that simply tests a few
commands for the SPI Flash. Other SPI connections are not tested for
now.

The SPI tests are broken down into 4 components:
  - Read the JEDEC ID from Flash over SPI and check it matches what we
  expect.
  - Erase a random sector (4 KiB) from flash over SPI and check that the
  entire sector was properly erased.
  - Write 10 random pages (256 B) to flash over SPI, and then read back
  the data to check that it matches what was written.
  - Write 1 random page, but with an 8x slower SPI clock speed.

The number of test iterations, and the number of sectors and pages
tested can be configured via compilation flags. These numbers are
purposefully kept low to keep runtime reasonable in Verilator.