Do not modify CS while SPI controller is active
Ensure that the CS is not changed whilst the SPI controller is active,
e.g. after a write into the FIFO which may not necessarily be completed
on the SPI bus, when running at slower SPI clock speeds.

It is safe to de-assert the CS line when a read operation has completed,
because that necessarily blocks until completion.

Co-authored-by: Adrian Lees <a.lees@lowrisc.org>