Add CHERI I2C Test
This commit adds an I2C test to the existing baremetal CHERI test
runner. For now, these are simple smoketests that test a few possible
connected I2C devices by reading out some information and checking it
matches what we expect. The tests are made based on the existing
`i2c_hat_id` demo being used as a non-CHERI check.

There are three tests, which can be configured based on which devices
are present:
  1. Read the Raspberry Pi HAT ID EEPROM header and check it.
  2. Read the Raspberry Pi HAT Accelerometer & Gyroscope and Magnetic
  Sensor `WHO_AM_I` registers and check these.
  3. Read the config and temperature value from the AS6212 Temperature
  sensor and check these values are sensible.

The number of test iterations, and the devices that are available can
be configured via compilation flags.