Add Sonata CHERI Pinmux Driver
This commit adds a simple Pinmux driver for Sonata. This is the version
that will hopefuly be upstreamed to `cheriot-rtos` in the future, but
with some minor changes so that it can run in a baremetal environment
(it has a constructor that takes a bounded capability, rather than using
the `MMIO_CAPABILITY` macro recommended to be used in the RTOS drivers).

This driver is being added here for usage in checks, and so that it can
live somewhere while there are potential software-breaking changes to
the pinmux RTL that can be tested and fixed before the driver is
upstreamed.

Sonata's pinmux only allows certain block inputs/outputs to be muxed
with certain pin inputs/outputs. For now, these combinations are only
explicit in the documentation. Instead, the driver provides a way to
check the valid range of indexes of selections for a given output pin or
block input, but users of the driver must check the documentation to see
which multiplexing options are available.