Add basic Pinmux CHERI check.
This commit adds a very basic pinmux sanity check for a CHERI execution
environment. Note that this is just a check and not a test. The check
simply logs to the UART, disables the UART0 pinmux to Serial 0 TX, logs
again, and then re-enables the UART0 pinmux to Serial 0 TX, and logs one
more time. If pinmux is working as expected, then when running the check
application you can see the first and third messages but not the second.

This is only intended to be a simple check, and not a test, and as such
does not test any other blocks or any more complex muxing functionality.