Double time for tests to run in simulation
Bumps the time available for CI tests to run in the Verilator simulation
environment from 30 to 60 seconds. Looking at recent CI runs containing
software changes with no caching, the `Nix Checks` step of the simulator
job appears to take around 20-30 seconds (given some variation), meaning
that soon we might get CI Failures as a result of timeout.

`nix flake check -L` will also fail locally more frequently on developer
machines due to this stringent timeout. This change hence allows tests
to be run locally, and gives a little more wiggle room for future tests
to be added.

This is required due to adding the e.g. hyperram and SPI tests which
take a bit of time to run in simulation.