Update CHERIOT RTOS sources to point to hyperram branch
To enable the CHERIoT RTOS test suite to function in CI, we need to
first update the CHERIoT RTOS source. It currently points to the main
branch of the lowrisc fork, but the version being used to build
software has since been moved to a branch containing a hyperram hack,
to allow the RTOS to continue to function with hardware changes.
Without this change, the CHERIoT RTOS test suite will not correctly run
on simulator in CI.

To keep sources in sync, this also updates the tag being used in the
Cmake build process to point to the same branch. This is what is being
tested in CI, and so (bar any unforeseen issues), this should have no
far-reaching effects.