Modify Pinmux driver names to make sense / be correct
I assumed that e.g. rph_g17 would map to the Raspberry Pi HAT header pin
17. This is not actually the case, there is no linear mapping. Instead
was necessary to consult the pin description and schematic.

This commit changes the enum output pin names to better reflect the
actual pin labelling on the chip, and also directly labels each pin
with its location on the board (e.g. P3.37, in the form header.number),
so that users of the driver can easily identify which pin it describes.
The existing hardware pin names are retained in the comment also, in
case they might be useful.