SPI CS control and interrupt logic
Modify the SPI controller to incorporate a parameterised number
of CS (Chip Select) lines to the attached peripherals.

Compatibility with the existing software tree is maintained by
combining the GPIO-driven CS lines with those from the controller
itself. In time the GPIO lines shall be dropped.

Generate FIFO-related and completion interrupts within the
SPI controller.