System info module with git info
This includes a git hash and a dirty bit.
The system information is generated using FuseSoC.
The block uses regtool for its register definitions.
The system information gets the information either via the git
repository it is in or via the Flake Nix environment if not.
- This likely breaks reproducibility/caching in many ways
Linux uses 48 bits to identify commits, so we choose to have two git
hash registers instead of just one, so we can encode 64 bits.

Co-authored-by: Harry Callahan <hcallahan@lowrisc.org>