Add Pinmux tests for I2C and SPI devices
This commit initialises the pinmux test suite for the Sonata board, and
adds two tests to this. This includes an I2C test and a SPI test. These
tests enable and disable the I2C (to the Raspberry Pi HAT) and SPI (to
the on-board SPI Flash) pins, and checks that corresponding I2C tests
(reading ID EEPROM, WHO_AM_I registers) and SPI tests (reading the JEDEC
ID from the Spi Flash) pass or fail correspondingly according to the
current pinmux configuration.

These tests can be seen as showcasing that basic pinmux selecting
functionality to enable/disable/re-enable pins is working as intended.