Add Pinmux GPIO and Muxing test
This commit adds two more tests to the Pinmux test suite: GPIO, and
Muxing.

The GPIO test is another single-device "callback" test. It takes a given
input pin and output in (configured as such), and then tests that
writing high/low states to the output pin can be read back from the
input pin after a short time. This is done a set number of times to
provide confidence that GPIO is muxed and working as expected. This
requires a manual connection between the pins on the board.

The Muxing test then combines the GPIO and UART tests, muxing two pins
between GPIO and UART functionality, and checking that the corresponding
GPIO read/write and UART send/receive tests pass/fail as would be
expected in the used Pinmux configurations. Where other tests simply
enabled/disabled pins via muxing, this actually switches between UART
and GPIO functionality on the same pins via muxing to give more
confidence in the Pinmux functionality. This again requires a manual
connection between the pins used on the board.