Add Pinmux UART test
This commit adds a UART test to pinmux, which tests using UART
functionality when the corresponding UART TX and/or RX pins are
enabled/disabled via Pinmux, ensuring that these tests correspondingly
fail/pass as would be expected under a working pinmux configuration.

The UART test is a single-device "callback" test - a cable is connected
on the board between the UART TX and RX output that are muxed, and the
test is then responsible for sending and receiving UART transmissions,
and verifying the content (randomly generated) matches.