Add pinmux testplan checker
This commit defines the structure of a manually-guided Pinmux testplan,
and adds a C++ file with a header which users can include to run their
own pinmux testplan. This is designed to be flexible and generic to
allow any pinmux configuration to be tested via appropriate tests.

Tests themselves can test GPIO (by connecting two pins in a loop,
writing to one and reading from the other), UART (by connecting two pins
in a loop, sending on one and receiving on the other), SPI (by reading
the JEDEC/Manufacturer ID of an attached PMOD SF3 N25Q256A Flash Memory
device) or I2C (by reading the JEDEC/Manufacturer ID of an attached
BH1745 Luminance and Colour Sensor Breakout, or a PMOD Colour).

Each test defines a test type, a set of pinmux output pins / block
assignments to make, the data for that test, whether any manual
intervention is required (and if so, an instruction), and the expected
result of that test (to allow for negative results).

Tests are then executed sequentially as defined in a testplan (an array
of tests). Depending upon the configured settings, failure of a test can
either instantly fail the entire testlpan, or progress as usual, or ask
the user whether they would like to retry the test (in case they carried
out the manual test wrong).