[sw] Rework i2c test to generally verify any sense HAT EEPROM
The EEPROM contains a UUID that is different for every board so a simple
binary comparison will only ever work for a specific board. Furthermore
there are a couple of sense HAT versions that cause further variations.

With this rework we instead read and verify the EEPROM contents by using
the CRCs that are part of the ID specification and check certain fields
(like the vendor string) are as expected.

This introduces some new utility functions for working with R-PI HAT
EEPROMs that may be useful for other Sonata software.