Add UART3 to Arduino Shield D0 & D1 via pinmux
This comit changes the pinmux RTL to additionally mux UART3 through the
Arduino shield output pins. This is done so that we have a pair of pins
that can be used for easily testing the pinmux muxing functionality; we
mux between the GPIO and UART RX/TX functionality of these pins to
perform GPIO and UART functionality testing with just the single device.

These pins are used instead of the equivalent RX/TX on the Raspberry Pi
Sense HAT Header because that is connected to the Rpi Sense HAT for I2C
testing and testing muxing of I2C pins. Without soldering, these RX/TX
pins cannot be easily re-used. Adding these pinmux pins instead enables
easier testing, and also exposes UART3 in more places to be used.