Update CHERIoT RTOS source for I2C driver updates
Updated from 34fc63bd6f11ba747b2e98e49f40eb5949809682
to c29bdeccc0335b81acea09c069fc94ca044a5a14, which just has two commits
to update the I2C driver to work with the latest vendored in block to
the RTL, and makes `blocking_write` check for failure so that we can
use it in tests to reliably check for failures (e.g. in the pinmux
test).