Add MicroSD pins to Pinmux Driver
This commit updates the Pinmux driver so that it is no longer out of
date with the most recent pinmux changes that introduce the MicroSD
card. This involves adding two more output pin registers for the MicroSD
clock & command pins, and allowing the Spi3 Receive to have 1 more value
selected (i.e. the MicroSD Dat0).