Raw SD block write/read test for manual use
Note that this manual test will corrupt the contents of
the card so it can only be run manually and will intentionally
insist that you wilfully insert an SD card before it performs
any writes to the card. This is to avoid inadvertent data loss.

SD card layer can perform initialisation, CID/CSD reading, and
single/multiple block read/write transfers from/to a number of
microSD cards used in testing. Only SDHC is supported and
3.3V cards only, but cards of 32GiB upwards from a number of
manufacturers have been tried.