Minor pinmux test improvements
Addresses some comments from the Pinmux tests PR after it was merged.
This adds in some missing SPI (re)-initialisation logic and does some
minor refactoring to make the code cleaner in places. It also makes a
minor change formatting the output of the USB test to make it more
consistent.