Bump CHERIoT RTOS to include updated GPIO driver
This bumps the version of CHERIoT RTOS used in both CMake and Nix, so
that we can access the newly updated GPIO driver. This GPIO driver
retains backwards compatibility with the previous interface but also
adds structs for operating with the other non-general GPIO blocks
(Raspberry Pi HAT, Arduino Shield and PMOD).