Make pinmux driver be generated by top_gen
Moves creation of the pinmux driver so that it is now automatically
generated by `top_gen`, so that there is no potential for possible
future issues with register / selector mismatch or mistakes from e.g.
typos as pinmux configuration changes are made.

This also updates the driver names to (for now, at least) only use
the pin / doc names and not use driver-unique names, as this was
judged to be beneficial (at least, for now) from discussion with others.