Update pinmux config to add SPI CS, and change PMOD SPI
Adds the SPI CS option to Pin 1 (i.e. 0) for PMOD0 and PMOD1. This
commit also changes PMOD0 to use SPI2 instead of SPI3, and PMOD1 to use
SPI3 instead of SPI4 (which will soon no longer exist after config
changes). This is to map the eventual desired pin mapping.