Update pinmux "all blocks" check to use CS & SPI-2 over PMOD
After changes to top-config, this commit does 2 things. Firstly, it
changes the SPI test in the pinmux checker to use the actual SPI chip
select now that it is available through pinmux, rather than just using a
GPIO for the same purpose.

Second, it updates the checker to use SPI2 and SPI3 instead of SPI3 and
SPI4, as this matches the desired mapping for pinmux in the end (no
SPI4). This migrates PMOD0 over to using SPI2 now so that these future
changes will not break this test.