Fix incorrect pinmux check pin logging
This was left over from an incorrect I2C pinmux config I did, where I
accidentally put the I2C pins on 1&2 instead of 3&4. The indexes are
still off by one because of the indexing change made to PMOD. This
commit simply updates the log output of the test to reflect the actual
pins that are being tested for I2C in the tests.