Minor refactor to pinmux checker to better use enums
Refactors the tests code to use the Spi, I2c and UART test enum values
to store the passed array offsets rather than relying on subtracting
some constant literal. This also adds a "NumX" to the end of each enum
so that anybody changing this code in the future can easily access the
number of devices without worries about that falling out of sync.