Make Spi Pinmux tests more resilient.
After doing any sort of pinmuxing with SPI, reset the SPI Flash using
the driver to ensure that e.g. muxing of the CS didn't place it in a
state that causes subsequent testing to become invalid.

Also replaces the use of the constant literal `3` with a
`sizeof(jedec_id)` as a good practice to remove constants from the code
and reduce the chance of things falling out of sync.