Pinmux doc names now omit index for blocks with 1 instance
For any blocks with only 1 instance (e.g. PWM), this omits the `[0]`
indexing bracket following the block name in any of the doc names, and
thus the block names. This means that these names then become something
like `pwm_out[0]` instead of `pwm[0].out[0]`, which is unecessarily
verbose and confusing when there is only one such instance.

Only affects the doc names, and nothing about the generated RTL.