Update software to match new GPIO device ranges
This commit updates the software that uses the full range of GPIO
devices (that is, the pinmux tests and the manual check procedure i.e.
the "all blocks" check). It splits the original definition of "PMOD" now
into "PMOD0" and "PMOD1", and correctly instantiates capabilities for
the new GPIO device MMIO ranges that matches the generated xbar
according to the updated definitions.