Update Pinmux checker to follow new GP_I pin mapping
The joystick has now moved to come after the User DIP switches in the
gpio_board input pin mapping. As such, the corresponding bit indexes in
`pinmux_checker` must be updated so that the manual pinmux check
procedure can continue to work.

This change also needs to be propagated to all other software in the
repository - this commit is just to get the manual pinmux checker
working.