Improve simulation match with FPGA
Re-express some wiring and ensure that inputs are defined
within top_verilator to achieve a better match with the
FPGA implementation in top_sonata.

In particular ensure that the output pins are driven high
when disabled as per the the FPGA builds, which is important
for SPI CS and UART TX lines.