Update common/legacy SW definitions
Updates the definitions of register locations/offsets/addresses, device
MMIO region bounds, the number of IRQs available, the number of each
device available, and the available GPIO registers for the common and
legacy software definition files.

Also adds a couple extra missing devices to the list of common
definitions (clint, ADC, revoker), and re-organises the file in order
of start address.