Update legacy I2C driver definition
The I2C block was updated, which now replaces the `Nack` interrupt with
a `controllerHalt` interrupt, which is asserted so long as there is a
bit set in the newly added `controllerEvents` register. This allows the
I2C block to provide multiple reasons for its controller FSM halting -
e.g. timeouts or loss of delegation, as well as NACKs.

This commit brings the legacy I2C driver up-to-date with the CHERIoT I2C
driver, adding in some extra checks on I2C writes for controller halts,
and making sure to clear the `controllerEvents` register when the
interrupt is raised.