Update bootloader to use latest GPIO drivers
Recent HW updates (i.e. the pinmux mapping) changes the GPIO register
definitions and mapping. This commit changes the use of `SonataGPIO` to
the more specific `SonataGpioBoard` from the updated upstream GPIO
drivers, and is kept separate from other changes so that this change to
the bootloader can easily be reviewed in isolation.