Update manual pinmux checker for new pin mapping changes
Updates the manual `pinmux_checker` application along with the
`pinmux_all_blocks_check` that uses it to match the new pin mapping
changes under Pinmux for 1.0.

This involves changing the pinmux assignment for a couple of pins to
match the changed order in the pinmux config changes. It also involves
reducing the number of UARTs that can be used by the the pinmux checker
from 1-2 down from 1-4, since the overall number of UARTs has also been
reduced by recent harwdare changes, and thus these non-existent UART
devices should not be exposed.

It finally also changes the number of SPIs exposed to just SPI0 and
SPI1 which are now exposed via Pinmux, and makes the necessary changes
to pinmux checker to facilitate this.