Implement other rendering orientations
The DPI model of the LCD panel now supports all rendering
orientations for other tests in simulation.

Modify the lcd_check code to support operation in landscape
orientation and not just portait. This was necessary to
assist with bring up of other simulation tests; 3 of the
possible 8 orientations are in use.

Eliminate the use of GPIO pins; LCD is now controlled using
the SPI controller Chip Selects and the final PWM output.