Add no_default_out option to pins in pinmux generator
When present and set to 'true' the default pinmux select for the pin
output will choose the first (constant) input. This means the pin will
be connected to no block initially.

Without this we had a situation where a block, such as a UART, outputted
on multiple pins by default with no way to change this behaviour.

This commit does not apply the option anywhere so the generated pinmux.
Additionall commits will be required to apply the option in appropriate
places.