Implement PWM tests for Sonata
Requires a jumper cable connecting the MikroBus PWM and PMOD0 pin 1 on
the board to pass. One test sets the period and duty cycles to a range
of different values and measures the PWM output via the GPIO to check
that these values are being correctly output, given a certain margin of
error due to the limited size of the counter.

The other test simply checks that when the period/counter is configured
to 0, the PWM always outputs 0 (low).

As part of this commit, also adds a definition for the PWM device to
`defs.h`.