Update SW definitions of number of GPIO/UART/SPI
Updates the definitions of the number of GPIO, UART and SPI devices that
are available in the software definitions file for Sonata to match the
changed nubmer of devices in the new pin mapping configuration. This
commit aims to keep changes minimal to get tests passing - it does not
look at solving any other issues with e.g. outdated definitions for now.