Update pinmux tests for mapping changes & convert SPI test to use external flash
Updates the pinmux tests for the changes introduced with the new pinmux
mapping. This primarily involves just changing the pins / devices used
to the devices that are now mapped on those pins, and updating
documentation to reflect this.

Also resets all pinmux logic after muxing to ensure that if more tests
are added afterwards or if the tests are run multiple times, that errors
are not introduced by the pinmux tests changing state.

Since the Application flash pinmux pins are no longer available through
pinmux, this also involves converting the SPI Pinmux test to change to
instead test using an extrenal SPI PMOD SF3 conecting to PMOD1. This
repurposes the existing test logic used by the pinmux checker to reduce
the amonut of additional code.