Attach Basic PMOD SF3 SPI DPI to PMOD1 in Verilator
Adds a basic PMOD SF3 DPI in Verilator, attached to the PMOD1 pins so
that the device can be used. This is modelled to be the same as the
existing Application Flash on Sonata, but with a different JEDEC ID that
it reports. Makes the Spi Flash take a JEDEC ID as a parameter to reduce
duplicated logic between these DPIs.

Disclaimer: I am not a hardware engineer, the hardware changes may
contain problems.