Add PWM Always High test & Misc. Improvements
Makes some small improvements to the PWM tests. First adds an "Always
high" PWM test, which checks that the PWM always outputs 1 when the duty
cycle is in some configuration greater where it is greater than the
period/counter. This acts as an opposite to the "Always Zero" test.

Second, this PR adds a couple more PWM loopback tests cases, to better
cover the range of duty cycles above 50%, which were previously not
being tested. These are 224/255 and 192/255.

Third, corrects a couple of typos.