Resurrect SD card operation
Combining CIPO lines directly without regard to Chip Selects
is risky, and in particular the line `appspi_d1` from the
flash is tri-stated without a pullup resistor on the FPGA
board.

Modifications to the top_verilator implementation to match
the top_sonata implementation more closely, and paving the
way for the introduction of the microSD card DPI model
(already built and operational).