Add timeout to plic test WFI & correctly fail on not firing interrupt
This commit makes sure that the `irq_fired` attribute is correctly reset
to `false` after each `wfi()` call in a test, which was previously
causing the test to pass even when the interrupt lines for a block were
not correctly hooked up in hardware.

As that would cause the entire test runner to halt, this commit then
replaces the ASM `wfi` functionality with a basic timeout for handling a
fired interrupt, which is set to 10 microseconds by default so as to not
cause lengthy waits when running in simulation. This timeout means that
the `wfi` function can actually terminate without an interrupt being
fired, and that the tests can now query the `irq_fired` line to
determine whether the test passed or failed, and log accordingly.