Insert delay on lcd_copi to avoid timing issue
Manually insert delay in the form of BUF (LUT1) cells directly before
the lcd_copi output buffer (OBUF) to satisfy the large hold requirement
of the LCD driver chip.

If left to add hold-fixing itself, the tool occasionally causes
setup timing failures on internal shift/loopback paths by placing the
hold-fixing (in the form of a hold detour) too early in the path.
Hold fixing as large as this should be put close to the port,
after all other paths have branch off. Why the tool decides to insert
such large delays so early in the path, I do not know.