Report correct GPIO num on sysinfo
The number of GPIO instances in the sysinfo block was being
under-reported by 1, because the dedicated board GPIO that is not
passing through pinmux was not counted in the Pinmux `GPIO_NUM`
parameter. This commit replaces instances of `GPIO_NUM+1` in the code
with a localparam `TotalGpioNum` in the same vein as the recent SPI
changes, and fixes this instance of undercounting the GPIO. This should
help prevent future issues from mismatches due to changes not being
synchronised elsewhere in the code.