Update SPI names/numbering in pin mapping diagram
In the latest RTL changes, `spi_board` was changed to go back through
Pinmux, and hence it became `spi0` again (this is the SPI for flash and
MicroSD). Hence SPI0 now becomes SPI1 and SPI1 becomes SPI2. This is
already up-to-date and working in `top_config.toml`, but until this
commit it had not yet been updated to reflect that in the pin mapping
diagram.