Change PMOD0&1 Pin names in mapping diagram to match top config
`top_config.toml` was already updated to change the names of
`PMOD{0,1}_{5,6,7,8}` to `PMOD{0,1}_{7,8,9,10}`. This commit updates the
pin mapping diagram to match the names of these pins accordingly. This
change was originally made so that pin names better reflect the
numbers/names given to pins by the Digilent specification, and not the
pin names on the schematic.

Also fixes a minor typo for `pmod1_4`, which said "sca" rather than
"sda" for the I2C1 Serial Data pin.