Update Raspberry Pi Header Pin names to remove SPI
The pin names for the Raspberry Pi HAT header's pins in
`top_config.toml` had already been stripped to remove references to
specific SPI pins (e.g. CE0, CE1, COPI, CIPO, SCK), because of the
inconsistencies between the usage of SPI pins on some devices, and to
avoid confusion between SPI0/SPI1 and the blocks that are referred to as
SPI0/SPI1 on our board (in reality, SPI0 on RPH uses our SPI1, and SPI1
on RPH maps to use our SPI2).

This commit just updates the pin mapping diagram to reflect the existing
changes to `top_config.toml` and the RTL.