Revocation control test revocation happens.
The test now writes two capabilities to the end of SRAM, one with base
zero and one with base at the end of SRAM. After the sweep it checks
that the base zero capability's tag stays in tact while the one at the
end of SRAM is cleared. Note this is different from the CLC test because
the revocation tags are cleared before testing so the CLC will not clear
the tags if the hardware revocer hasn't already.