Fix only PWM 0 being enabled via Pinmux
Currently, all the PWMs (minus the LCD backlight) are being routed
through pinmux, but the value being passed for the PWM output
enable is the constant `'{'b1}'`. Because the inner value is `'b1`,
this is extended to fill the 7 bit PWM output range as `000001`,
instead of using `'1` which would extend to `111111`.

This commit makes the casting/extension of these constant enable
values more explicitly match the PWM output size and module count, to
ensure that all outputs of all PWM devices are being enabled through
pinmux.