Add `pinmux_pwm_test` to test runner
This commit introduces a new `pinmux_pwm_test` that will run in CI,
which tests that PWM appears to be working correctly when it is muxed
over PMOD, and that it doesn't work when disabled via Pinmux.

This re-uses the loopback testing code from the PWM tests that reads the
PWM output via GPIO to determine the duty cycle and period, and use that
to verify that PWM operation is as expected. There is some small
refactoring to make that possible, including a fix in which the PWM
loopback test was accidentally using hardcoded GPIO 0 instead of the
specified pin constant (which was also 0, hence no previous problems).

Requires an additional jumper cable to FPGAs to pass the test, which can
be optionally disabled by the same `PINMUX_CABLE_CONNECTIONS_AVAILABLE`
method as for tests where this was previously required.