xbar: debug host access to hyperram, system_info
With the RTOS mainline gaining some support for code in HyperRAM, it
would be nice to be able to load such images in over OpenOCD without
needing to write code to do bounce buffering.

While here, also give the debug host access to the system_info
peripheral, to provide some more diagnostic information.