Read from LCD to determine if it needs an offset
Some LCDs on some more recent boards have the GM pads of the internal
ST7735 controller incorrectly set for the actual resolution of the
panel included in the assembly, resulting in an image offset.
We can detect this by observing side-effects of this misconfiguration
by writing and reading to the internal frame buffer.
For the mis-configured case, we add a one or two pixel offset to the
X/Y coordinates when we write pixel data to the LCD.

Written in a manner to increase portability.