Add Sonata One-compatible Sonata XL build target
Add Sonata XL bitstream build target that maintains
software compatibility with the Sonata One board.
Tweaks the design using `ifdef` statements and alternative XDC files.
Uses the extra block RAM of the Sonata XL FPGA to create a faux-HyperRAM
(1 MiB only) so that the memory map can be kept the same.
Does not yet make use of the extra board-to-board connectors
present only on the XL board.

Change the fusesoc target from `synth` to `synth_xl` to build.