Use actual widths for GPIO instances to save area
Better to instantiate each gpio_core with the correct in/out widths
than to use widths of 32 for all .
This avoids redundant logic being included in the bitstream,
such as the input sunchronisers which the tool is prevented from
optimising away.
Also the PCINT logic which is sensitive to constant-zero
signals in one mode, and so cannot be optimised away if the
unconnected bits are assumed to be constant-zero.

Effect on utilisation:
Slice LUTs: 83.29% -> 80.90% (-2.39%pt)
Slice Registers: 25.37% -> 23.95% (-1.42%pt)