Narrow GPIO `RegAddrWidth` a bit
Narrow the `RegAddrWidth` GPIO module parameter from 12 to 9 bits.
Should reduce the number of bits that need to be compared in
`gpio_array` to determine which instance to select.
Probably a very minor area saving, but every little counts.

Also update the default `RegAddr` in `gpio_core` to be more realistic.