Make the idle state of the COPI line configurable
The SD card interface is specified on the assumption on that the
COPI line remains high when the SPI controller is not transmitting.
Make the inactive/idle state of the COPI line configurable, but
keep the default as zero to avoid inadvertently changing the
behaviour of other SPI interfaces.

This change permits the SD card driver to be simplified and the
CPU load to be reduced, because presently the driver must keep
the TX FIFO populated with bytes having the value 0xff whenever
data is being read.