Reduce access latency of GPIO blocks
Rework the `gpio_core` to provide response data in the same cycle
as the request to reduce the access latency.
This also allows the rvalid logic to be removed as the rdata
can be selected using the still-valid request address.

Reduce the value of the `AccessLatency` parameter of the TL-UL adapter
used to connect the GPIO blocks to the system bus,
to match the `gpio_core` read latency changes.