Improve the descriptions of the HyperRAM-related switches
There are two switches and neither is exclusively HyperRAM-related.

One controls the clocking scheme and gives a more accurate
simulation of the USB device and/or HyperRAM controller as desired.

The second controls whether a SRAM model is used in place of the
HyperRAM controller; this may be used in simulation for a faster
simulation or in synthesis when targeting SonataXL which has no
HyperRAM.