Revert HyperRAM clock frequency to 100MHz
There is a suspected physical-level issue with recent builds
which may be related to the increased clock frequency. Although
that is uncertain, there are two other reasons to propose
reverting the commit that increased the frequency:

1. There is a throughput mismatch between the HBMC read speed
and the Sonata system with only a 32-bit Upstream FIFO. This
could maybe lead to data loss because there is no braking/
backpressure mechanism. It was intended to widen `ufifo` to
64 bits but this has not yet been completed and proven.

2. The increased frequency offers a performance uplift of only
about 10-16% with the current design. A greater boot should be
possible later but that requires some rearchitecting.